Moving from classes to modules
